requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
transformers==4.36.2
torch==2.1.2
pillow==10.2.0
gTTS==2.5.1
pydub==0.25.1
SpeechRecognition==3.10.1
pyrebase4==4.7.1
python-dotenv==1.0.1